+++ src/ska_tmc_dishleafnode/commands/configure_band_command.py
@@ -33,10 +33,12 @@ class ConfigureBand(DishLNCommand):
         op_state_model,
         adapter_factory=None,
         logger: logging.Logger = LOGGER,
+        is_configure_command: bool = False,
     ):
         super().__init__(
             component_manager, op_state_model, adapter_factory, logger
         )
+        self.is_configure_command = is_configure_command
 
     # pylint: disable=unused-argument
     def configure_band(
@@ -63,7 +65,7 @@ class ConfigureBand(DishLNCommand):
         # Indicate that the task has started
         self.task_callback = task_callback
         self.task_callback(status=TaskStatus.IN_PROGRESS)
-        if self.component_manager.is_configure_command is False:
+        if self.is_configure_command is False:
             self.set_command_id(__class__.__name__)
             self.component_manager.start_timer(
                 self.timeout_id,
@@ -83,10 +85,10 @@ class ConfigureBand(DishLNCommand):
                 result=(ResultCode.FAILED, message), exception=message
             )
         else:
-            if self.component_manager.is_configure_command is False:
+            if self.is_configure_command is False:
                 self.start_tracker_thread(
-                    "get_dish_configured_band",
-                    [argin],
+                    "get_configure_band_result_code",
+                    [ResultCode.OK],
                     task_abort_event,
                     self.timeout_id,
                     self.timeout_callback,